requirements.txt
Flask==2.0.1
Flask-WTF==0.15.1
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.1
gunicorn==21.2.0
pytest
werkzeug==2.0.1
psycopg2
mysql-connector-python